pyproject.toml
[project]
name = "stanley-okwii"
version = "0.1.0"
description = "Scripts that render the CV and fetch stats for stanleeokwii.info"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "rendercv[full]>=2.8",
    "pyyaml>=6.0",
]
